fix(takum): report radix-2 exponent bounds, widen takum_log::scale()
Three defects from review round 2 of #1304.

numeric_limits exponent bounds were not radix-2 exponents.  Both
specializations declare radix == 2, but reported a characteristic instead,
and without the C++ convention that min_exponent is one MORE than the
smallest normalized power.  For takum_log the error compounded: its
characteristic counts powers of sqrt(e), so the reported range overstated
the true one by 1/log2(sqrt(e)) ~ 1.39x, and min_exponent10 then applied
log10(sqrt(e)) on top, converting the base a second time.

The bounds now come from the extreme ENCODINGS, magnitude 1 and
magnitude_mask(), rather than from min/max_characteristic().  Those are the
range the format advertises, which the extremes need not attain: takum<12,3>
has no trailing field in its narrowest DR, so minpos sits at c == -254 while
the advertised minimum is -255.  All 12 configurations whose extremes fit in
a double now agree exactly with std::ilogb(minpos) + 1 and
std::ilogb(maxpos) + 1, for both variants.  cfloat's specialization documents
the same +1 convention.

takum_log::scale() had undefined behaviour.  It narrowed to int, but at
rbits == 5 the characteristic reaches ~2^32, so the value exceeds int range.
UBSan reported "signed integer overflow" and takum_log<16,5> gave maxpos
scale INT_MIN and minpos INT_MAX -- an inverted range.  It now returns
int64_t, which is also what takum<>::scale() returns, so the conversion is
removed rather than clamped.  Reachable through manipulators.hpp's
components(), which accepts either variant.

VerifyTruncatedRounding silently skipped a non-ok encode status.  c is range
checked and k+1 still fits the stored field, so saturation there is a defect;
report it instead of continuing, which would otherwise let a regression pass
while other cases kept `exercised` non-zero.

An earlier comment credited attributes.hpp's minpos_scale()/maxpos_scale()
as the path that reaches takum_log::scale().  That was wrong -- those
templates only accept takum<> -- and is corrected to manipulators.hpp.

+++ include/sw/universal/number/takum/takum_log_impl.hpp
@@ -321,6 +321,13 @@ public:
 	// that scale() is base 2 for every number system.  The characteristic is in
 	// units of sqrt(e), so it is converted here rather than returned directly.
 	//
+	// Returns int64_t, as takum<>::scale() does.  An int cannot hold the result: at
+	// rbits = 5 the characteristic reaches ~2^32, so takum_log<16,5> maxpos scales to
+	// ~3.1e9 and the narrowing conversion is undefined behaviour -- UBSan reported
+	// maxpos as INT_MIN and minpos as INT_MAX, an inverted range.  Widening the return
+	// type removes the conversion rather than clamping it, and manipulators.hpp's
+	// components() reaches this for every takum_log configuration.
+	//
 	// Accuracy: sqrt(e)^l is a power of two only for l = 2k ln2, which is irrational
 	// for k != 0, so no encoding sits exactly on a boundary and the floor below is
 	// always well defined.  It is computed in double, though, which at |l| ~ 255
@@ -330,12 +337,11 @@ public:
 	// caller ever needs it.  Note that std::ilogb(double(*this)) is NOT the fix: it
 	// agrees with this on every uniformly sampled encoding, is no better on the
 	// boundary set (both are coin flips there), and returns INT_MAX for the ~34% of
-	// wide-rbits encodings whose magnitude overflows a double -- which is precisely
-	// where attributes.hpp's minpos_scale() / maxpos_scale() land.
-	CONSTEXPRESSION int scale() const noexcept {
+	// wide-rbits encodings whose magnitude overflows a double.
+	CONSTEXPRESSION int64_t scale() const noexcept {
 		if (iszero() || isnar()) return 0;
-		double log2_magnitude = logarithmic_value() * log2_of_base;
-		int    s = static_cast<int>(log2_magnitude);
+		double  log2_magnitude = logarithmic_value() * log2_of_base;
+		int64_t s = static_cast<int64_t>(log2_magnitude);
 		if (log2_magnitude < 0.0 && static_cast<double>(s) != log2_magnitude) --s;  // floor
 		return s;
 	}
